refactor: removed legendType and applyToOption variable in VectorLayer
