+++ sites/geohub/src/lib/helper/index.ts
@@ -42,4 +42,5 @@ export * from './isStyleChanged';
 export * from './isInt';
 export * from './isRgbRaster';
 export * from './isUniqueValueRaster';
+export * from './isVectorIntervalExpression';
 export * from './generateHashKey';
